==> 3_merging_Num&Cat_features.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)
#pd.set_option('display.max_rows',None)

# Reading the CSV File
bank_df = pd.read_csv('bank_additional_full.csv',sep=';',na_values='unknown')

# Getting the dataset info:
bank_df.info()

## Processing the Object/Categorical Dataset

# Segregating the Object/Categorical columns
cat_df = bank_df.select_dtypes(include='object')

# Getting the dataset info.
cat_df.info()

# Getting rid of the response variable y from the categorical Dataframe cat_df
cat_df = cat_df.drop(columns='y')

# Segregating the categorical columns having null values
cat_df_null = cat_df.loc[:,cat_df.isnull().any()]

# Visualizig the above dataframe
cat_df_null.info()

# Segregating the categorical columns having No Null values
cat_df_notnull = cat_df.loc[:,~cat_df.isnull().any()]

# Visualizing the above dataframe
cat_df_notnull.info()

#### Creating the dummies variables for categorical columns without missing values

# Getting the Dummies
cat_df_notnull = pd.get_dummies(cat_df_notnull.astype('category'),drop_first=True,prefix=cat_df_notnull.columns)

cat_df_notnull.info()

#### Getting the dummies for categorical columns with missing values

# Getting the info of the object/categorical columns having missing values.
cat_df_null.info()

# Getting the dummies (without dropping the first column using drop_first=True as NaN columns are automatically dropped)
cat_df_null = pd.get_dummies(cat_df_null.astype('category'),prefix=cat_df_null.columns)

# Visulazing the resulting dataset
cat_df_null.info()

# Standardizing the column names of the new dummied dataframe
cat_df_null.columns = cat_df_null.columns.str.strip().str.replace('.','_')

# Correcting a single column name from job_admin. to job_admin
cat_df_null.rename(columns = {'job_admin_':'job_admin'}, inplace = True) 

cat_df_null.info()

### Concatenating the above 2 categorical dataframes to produce a single one.

# Concatenating cat_df_notnull and cat_df_null to produce final categorical dataframe cat_df_final
cat_df_final = pd.concat([cat_df_notnull,cat_df_null],axis=1)

cat_df_final.info()

### Merging both Numerical & Categorical columns to create a final Dataset.

# Reading the Numerical Column Dataset.
num_df = pd.read_csv('Num_df.csv')

# Checking the Numerical Datset
num_df.info()

# Concatenating both Numerical as well as Categorical Dataset to produce a  Final Dataset
final_df = pd.concat([num_df,cat_df_final],axis=1)

# Checking the final dataset
final_df.info()

# Saving the Final dataset as the CSV file.
final_df.to_csv('Final_df.csv',index=False)

# Saving the Final Categorical dataframe as CSV.
cat_df_final.to_csv('Cat_df.csv',index=False)
