==> 5_a_Modeling.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)

# Loading both versions of Training sets (Full Feature & Reduced Feature)
X_train_red = pd.read_csv('X_train_final.csv')
X_train = pd.read_csv('X_train_full_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading both versions of Test sets (Full Feature & Reduced Feature)
X_test_red = pd.read_csv('X_test_final.csv')
X_test = pd.read_csv('X_test_full_final.csv')
y_test = pd.read_csv('y_test.final.csv')

### As per the Dataset Metadata , the column duration might leak information, since if duration = 0 then y=0 almost surely and after the end of the call, y is more or less known. So in order to have a realistic predictive model , the column 'duration' should be discarded. 

# Displaying the head of the reduced feature training dataset
X_train_red.head()

#### Deleting  'duration' feature from the Feature space of both training and test datsets.

#### Deleting 'duration' column from all the Training Dataset.
del X_train['duration']
del X_train_red['duration']
del X_test['duration']
del X_test_red['duration']

#### Saving the Resulting datasets back as CSV files

X_train.to_csv('X_train_full_final.csv',index=False)
X_train_red.to_csv('X_train_final.csv',index=False)
X_test.to_csv('X_test_full_final.csv',index=False)
X_test_red.to_csv('X_test_final.csv',index=False)

## Baseline Model: Dummy Classifier with Default Parameters 

# Importing Dummy classifier from Scikit Learn
from sklearn.dummy import DummyClassifier

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Instantiating the Dummy classifier object
bm = DummyClassifier(random_state=42)

#### Fitting dummy classifier (baseline model) to both full_feature training & test  sets and computing resulting roc_auc scores.

# Fitting the  dummy classifier to Training set with all features.
bm.fit(X_train,y_train)

# Accuracy for the full feature Training set with baseline model
bm.score(X_train,y_train)

# Accuracy for the full feature Test set with baseline model
bm.score(X_test,y_test)

# Predicting the probabilities of y=1 for the full feature training set
y_pred_train = bm.predict_proba(X_train)

print('The ROC AUC for the full_feature Training set is:',roc_auc_score(y_train,y_pred_train[:,1]))

# Predicting the probabilities of y=1 for the full feature test set
y_pred = bm.predict_proba(X_test)

print('The ROC AUC for the full_feature test set is:',roc_auc_score(y_test,y_pred[:,1]))

#### Fitting dummy classifier (baseline model) to both Reduced_feature training & test  sets and computing resulting roc_auc_scores.

# Fitting dummy classifier to the reduced feature Training set.
bm.fit(X_train_red,y_train)

# Accuracy for the reduced feature Training set with baseline model.
bm.score(X_train_red,y_train)

# Accuracy for the reduced feature Test set with baseline model.
bm.score(X_test_red,y_test)

# Predicting the probabilities of y=1 for the reduced feature training set
y_pred_red_train = bm.predict_proba(X_train_red)

print('The ROC AUC for the reduced feature training set is:',roc_auc_score(y_train,y_pred_red_train[:,1]))

# Predicting the probabilities of y=1 for the reduced feature test set
y_pred_red = bm.predict_proba(X_test_red)

print('The ROC AUC for the reduced_feature test set is:',roc_auc_score(y_test,y_pred_red[:,1]))

### Observations:
### 1) The roc_auc score for the reduced feature test set is  equal to that of full feature test set, indicating the presence of noise features in the full feature set.
### 2) The training set  roc_auc scores for both full feature & reduced feature training sets are less than that of corresponding test sets, probably due to underfitting.
### 3) The test set accuracy for both full feature & reduced feature test sets are less than 88.734%, which could be achieved by labeling all the test set  instances (y) equal to 0, the % of class 0 (majority class) in the whole dataset. Thus the default dummy classifier is certainly underfitting the training set.

## Model_1: Logistics Regression with Tuned Hyperparameters using Optuna.

# Importing  hyperparamater tuning optimizer optuna
import optuna

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score
from sklearn.linear_model import LogisticRegression

# Defining the appropriate objective function for the Logistic regression classifier
def objective_wrappper_ls(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Logistics Regression parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'C': trial.suggest_loguniform('C', 1e-5, 1e2),
        'l1_ratio':trial.suggest_uniform('l1_ratio',0,1),
        'class_weight':trial.suggest_categorical('class_weight',cl_weight),
        'penalty':trial.suggest_categorical('penalty',['none','elasticnet'])
            
        }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def train_test_roc_auc(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Instantiating the logistic Regression classifier
lr_s =  LogisticRegression(random_state=42,solver='saga',n_jobs=5)

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(10,random_state=42)

# Defing the class weights
cl_weight = [None,'balanced',{0:1.0,1:9.0},{0:1.0,1:10},{0:1.0,1:11},{0:1.0,1:12}]

#### Computing the best hyperparameters for the  logistic Regression using full feature Training Set.

# Extracting the best model parameters and best study score
best_study_score, best_study_params = train_test_roc_auc(X_train, y_train, lr_s, objective_wrappper_ls, cv_strat, n_trials=200)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ',best_study_params))

#### Computing the  full feature roc_auc score for the test data using the best study Parameters

# Obtaining the best full feature LR model by setting best study parameters.
lr_f = lr_s.set_params(**best_study_params)

# fitting the best Logistics regression model on the full feature training set
lr_f.fit(X_train,y_train)

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} is '.format(f_set,t_set,model_name),roc_auc_score(y,y_pred[:,1]))

# Calculating the full feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train, y_train, lr_f, 'full feature', 'training', 'Logistic Regression')

# Calculating the full feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test, y_test, lr_f, 'full feature', 'test', 'Logistic Regression')

print('The accuracy for the full feature test set is: ',lr_f.score(X_test,y_test))

# Saving the full feature best Logistic Regression model 
import joblib
joblib.dump(lr_f,'Log_Reg_Full.joblib')

#### Computing the best hyperparameters for the  logistic Regression using Reduced feature Training Set.

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, lr_s, objective_wrappper_ls, cv_strat, n_trials=200)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ',best_study_params))

#### Computing the  Reduced feature roc_auc score for the test data using the best study Parameters

# Obtaining the best reduced feature LR model by setting best study parameters.
lr_R = lr_s.set_params(**best_study_params)

# fitting the best Logistics regression model on the reduced feature training set
lr_R.fit(X_train_red, y_train)

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, lr_R,'reduced feature', 'training', 'Logistic Regression')

# Calculating the reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, lr_R, 'reduced feature','test','Logistic Regression')

# Computing the accuracy score for the Reduced feature test set 
print('The accuracy for the reduced feature test is: ',lr_R.score(X_test_red,y_test))

# Saving the Reduced feature best Logistic Regression model 
joblib.dump(lr_R,'Log_Reg_Reduced.joblib')

## Observations:
### 1) From the above analysis we can clearly see that roc_auc score of the full feature test set using tuned Logistic regression model is almost equal to that of  corresponding reduced feature test set , confirming the earlier suspicion that there are lot of noise features in the original dataset.
### 2) The roc_auc test set score for the tuned Logistic models are much higher than the scores from the corresponding  baseline models, which was expected.
### 3) The roc_auc test scores are more than their training counterparts for both full feature & reduced feature datasets ,perhaps due to underfitting. May be a more flexible model will yield better result.

## Defining  Reward_Risk Ratio for a Family of Machine Learning Models:

###  R_R Ratio = _Mean of  K Fold CV Score Metric of Training Data / Std_Dev of K Fold CV Score Metric of Training           Data_

### _The R_R ratio may be helpful in choosing among models having approximately same computational complexity or from the same family._

### Calculating R_R ratio for best Logistic Regression Model 

# Loading the best logistic regression Model
import joblib
lr_R = joblib.load('Log_Reg_Reduced.joblib')

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(10,random_state=42)

# Computing the Reward, Risk of the Logistic Regression Model
score_Log_Reg = cross_val_score(lr_R, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best Logistics Regression Model using roc_auc metric is: ',np.mean(score_Log_Reg))

print('The risk associated with the best Logistics Regression Model using roc_auc metric is: ',np.std(score_Log_Reg))

# Computing the Reward of the best Logistic Regression Model
R_R_Ratio_Log_Reg = np.mean(score_Log_Reg)/np.std(score_Log_Reg)

print('The reward risk  ratio for the best Logistics Regression Model using roc_auc metric is: ',R_R_Ratio_Log_Reg)

### R_R Ratio for the best Logistic Regression using reduced feature set is: 49.17219276619884
