==> 6_a_Modeling.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)

# Loading both versions of Training sets
X_train_red = pd.read_csv('X_train_final.csv')
X_train = pd.read_csv('X_train_full_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading both versions of Test sets
X_test_red = pd.read_csv('X_test_final.csv')
X_test = pd.read_csv('X_test_full_final.csv')
y_test = pd.read_csv('y_test.final.csv')

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Importing  hyperparamater tuning optimizer optuna
import optuna

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(10,random_state=42)

 ## Model_2: Random Forest Classifier with Tuned Hyperparameters using Optuna.

# Defining the class weights
cl_weight = [None,'balanced',{0:1.0,1:9.0},{0:1.0,1:10},{0:1.0,1:11},{0:1.0,1:12}]

# Defining the appropriate objective function for the Random Forest classifier
def objective_wrappper_rf(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Random Forest parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'max_depth': trial.suggest_categorical('max_depth',list(range(2,50))+ [None]),
        'n_estimators':trial.suggest_int('n_estimators',100,2000,10),
        'class_weight':trial.suggest_categorical('class_weight',cl_weight),
        'min_samples_leaf':trial.suggest_loguniform('min_samples_leaf',.00001,.1)
          }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def train_test_roc_auc(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Importing Random Forest Classifier from  Sklearn
from sklearn.ensemble import RandomForestClassifier

# Instantiating the Random forest classifier
rf_s = RandomForestClassifier(n_jobs=5, random_state=42)

#### Computing the best hyperparameters for the Random Forest classifier using full feature Training Set.

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train, y_train, rf_s, objective_wrappper_rf, cv_strat, n_trials=100)

print('The best roc_auc_score for the study is: ', best_study_score)

print(('The best study parameters for the classifier are: ', best_study_params))

#### Computing the  full feature roc_auc score for the test data using the best study Parameters

# Obtaining the best full feature RF model by setting best study parameters.
rf_f = rf_s.set_params(**best_study_params)

# fitting the best Random Forest model on the full feature training set
rf_f.fit(X_train, y_train)

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} classifier is '.format(f_set, t_set, model_name), roc_auc_score(y,y_pred[:,1]))

# Calculating the full feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train, y_train, rf_f, 'full feature', 'training', 'Random Forest')

# Calculating the full feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test, y_test, rf_f, 'full feature', 'test', 'Random Forest')

print('The accuracy for the full feature test set is: ',rf_f.score(X_test, y_test))

# Saving the full feature set best Random forest Classifier 
import joblib
joblib.dump(rf_f,'Rand_Forest_Full.joblib')

#### Computing the best hyperparameters for the  Random Forest Classifier using Reduced feature Training Set.

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, rf_s, objective_wrappper_rf, cv_strat, n_trials=100)

print('The best roc_auc_score for the study is: ', best_study_score)

print(('The best study parameters for the classifier are: ', best_study_params))

#### Computing the  reduced feature roc_auc score for the test data using the best study Parameters

# Obtaining the best reduced feature RF model by setting best study parameters.
rf_R = rf_s.set_params(**best_study_params)

# fitting the best Random Forest model on the reduced feature training set
rf_R.fit(X_train_red, y_train)

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, rf_R, 'reduced feature', 'training', 'Random Forest')

# Calculating the reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, rf_R, 'reduced feature', 'test', 'Random Forest')

# Saving the reduced feature set best Random forest Classifier 
joblib.dump(rf_R,'Rand_Forest_Reduced.joblib')

### Calculating R_R ratio for best Random forest Classifier.

# Loading the best Random forest Classifier model
import joblib
rf_R = joblib.load('Rand_Forest_Reduced.joblib')

# Computing the CV scores using sklearn's cross_val_score
score_Rand_forest = cross_val_score(rf_R, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best Random Forest Classifier using roc_auc metric is: ',np.mean(score_Rand_forest))

print('The risk associated with the best Random Forest Classifier using roc_auc metric is: ',np.std(score_Rand_forest))

R_R_Ratio_Rand_forest = np.mean(score_Rand_forest)/np.std(score_Rand_forest)

print('The reward risk ratio for the best Random Forest Classifier using roc_auc metric is: ',R_R_Ratio_Rand_forest)

#### The R_R ratio for the best Random Forest Classifier using roc_auc metric is:  52.904437509475876

## Observations:
### 1) The above results clearly show that the full feature dataset contains noisy components. Thus we got better roc_auc for the reduced feature test set than full feature test set. Further the amount of overfitting reduced  when reduced feature test set was used. _Hence we would only use reduced feature set from now on._
### 2) The R_R Ratio as well as test set roc_auc for the best Random Forest classifier are more than those of corresponding Logistic Regression classifier, thus former  fits the dataset better than the latter.
### 3) Random forest classifier overfitting is generally due to high test set variance. Perhaps using extra trees classifier, which trades off high variance for higher bias, may help.

## Model_3: Extra Trees Classifier with Tuned Hyperparameters using Optuna.

# Defining the appropriate objective function for the Extra Trees classifier
def objective_wrappper_ext(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Extra Trees parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'max_depth': trial.suggest_categorical('max_depth',list(range(2,50))+ [None]),
        'n_estimators':trial.suggest_int('n_estimators',100,2000,10),
        'class_weight':trial.suggest_categorical('class_weight',cl_weight),
        'min_samples_leaf':trial.suggest_loguniform('min_samples_leaf',.00001,.1)
          }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Importing Extra Trees Classifier from  Sklearn
from sklearn.ensemble import ExtraTreesClassifier

# Instantiating the Extra Trees classifier
ext_s = ExtraTreesClassifier(n_jobs=5,random_state=42)

#### Computing the best hyperparameters for the Extra Trees Classifier using Reduced feature Training Set.

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, ext_s, objective_wrappper_ext, cv_strat, n_trials=100)

print('The best roc_auc_score for the study is: ', best_study_score)

print('The best study parameters for the classifier are: ', best_study_params)

#### Computing the  reduced feature roc_auc score for the test data using the best study Parameters

# Obtaining the best reduced feature Extra Trees model by setting best study parameters.
extr_R = ext_s.set_params(**best_study_params)

# fitting the best Extra Trees model on the reduced feature training set
extr_R.fit(X_train_red, y_train)

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, extr_R, 'reduced feature', 'training', 'Extra Trees')

# Calculating the reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, extr_R, 'reduced feature', 'test', 'Extra Trees')

# Saving the reduced feature set best Extra Trees Classifier  
joblib.dump(extr_R, 'Extra_Trees_Reduced.joblib')

## Observations:
### 1) From the above analysis, we can clearly see that Extra Trees classifier did a good job in reducing overfitting to a large extent, probably by reducing test set variance.
### 2) Though overfitting was reduced, test set roc_auc score also suffered, probably due to corresponding increase in test set bias, which undermined any corresponding decrease in variance.

### Trying bias reduction in Extra Trees by further further tuning  class weight parameter & n_estimators.

# Defing the new class weights
cl_weight_extra = ['balanced',{0:1.0,1:10},{0:1.0,1:12},{0:1.0,1:13},{0:1.0,1:14},{0:1.0,1:15},{0:1.0,1:16},
                  {0:1.0,1:17},{0:1.0,1:18},{0:1.0,1:19},{0:1.0,1:20}]

# Defining the appropriate objective function for the Extra Trees classifier
def objective_wrappper_ext_1(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Extra Trees parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'max_depth': trial.suggest_categorical('max_depth',list(range(5,50))+ [None]),
        'n_estimators':trial.suggest_int('n_estimators',1500,2000,10),
        'class_weight':trial.suggest_categorical('class_weight',cl_weight_extra),
        'min_samples_leaf':trial.suggest_loguniform('min_samples_leaf',.00001,.1)
          }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, ext_s, objective_wrappper_ext_1, cv_strat, n_trials=100)

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ', best_study_params)

#### Computing the  reduced feature roc_auc score for the test data using the fine tuned best study Parameters

# Obtaining the best reduced feature Extra Trees model by setting tuned best study parameters.
extr_R_tu = ext_s.set_params(**best_study_params)

# fitting the best tuned Extra Trees model on the reduced feature training set
extr_R_tu.fit(X_train_red, y_train)

# Calculating the reduced feature training set roc_auc score using the tuned best study parameters
cal_roc_auc(X_train_red, y_train, extr_R_tu, 'reduced feature', 'training', 'tuned Extra Trees')

# Calculating the reduced feature test set roc_auc score using the tuned best study parameters
cal_roc_auc(X_test_red, y_test, extr_R_tu, 'reduced feature', 'test', 'tuned Extra Trees')

# Saving the reduced feature set best Extra Trees Classifier  
joblib.dump(extr_R_tu,'Extra_Trees_Reduced_Tuned.joblib')

## Observations:
### 1) The tuned Extra Tree Classifier has further reduced the overfitting by decreasing the gap between the training set & test set roc_auc score. However the test set roc_auc has also decreased by a small amount.


### Calculating R_R ratio for best Extra Trees Classifier .

# Loading the best Extra Trees Classifier model
extr_R = joblib.load('Extra_Trees_Reduced.joblib')

# Computing the CV scores using sklearn's cross_val_score
score_Extra_Trees = cross_val_score(extr_R, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best Extra Trees Classifier using roc_auc metric is: ',np.mean(score_Extra_Trees))

print('The risk associated with the best Extra Trees Classifier using roc_auc metric is: ',np.std(score_Extra_Trees))

R_R_Ratio_Extra_Trees = np.mean(score_Extra_Trees)/np.std(score_Extra_Trees)

print('The reward risk ratio for the best Extra trees Classifier using roc_auc metric is: ',R_R_Ratio_Extra_Trees)

#### The R_R ratio for the best Extra trees Classifier using roc_auc metric is:  52.1195604201456

### Calculating R_R ratio for tuned Extra Trees Classifier model

# Loading the best tuned Extra Trees Classifier model
extr_R_tu = joblib.load('Extra_Trees_Reduced_Tuned.joblib')

# Computing the CV scores using sklearn's cross_val_score
score_Extra_Trees_tuned = cross_val_score(extr_R_tu, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the tuned Extra Trees Classifier using roc_auc metric is: ',np.mean(score_Extra_Trees_tuned))

print('The risk associated with the tuned Extra Trees Classifier using roc_auc metric is: ',np.std(score_Extra_Trees_tuned))

R_R_Ratio_Extra_Trees_tuned = np.mean(score_Extra_Trees_tuned)/np.std(score_Extra_Trees_tuned)

print('The reward risk ratio for the tuned Extra trees Classifier using roc_auc metric is: ',R_R_Ratio_Extra_Trees_tuned)

#### The R_R ratio for the best tuned Extra trees Classifier using roc_auc metric is:  53.373843721800746

## Observations:
### 1) The tuned extra trees classifier has the best R_R ratio of all the tree based classifiers trained on the reduced feature dataset , due to low std. dev. of the cross validated roc_auc scores. Hence based upon this metric the best choice among the Tree based model is Tuned Extra Trees Classifier.

### R_R Ratio for the best Tree based bagging classifier using roc_auc score is:  53.373843721800746,  corresponding to Tuned Extra Trees classifier.

### _However keeping everything into account, for this dataset, the best Tree based bagging classifier is Optuna tuned Random Forest Classifier._
