==> 6.1_a_Modeling.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)

# Loading Training set
X_train_red = pd.read_csv('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading both versions of Test sets
X_test_red = pd.read_csv('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Importing  hyperparamater tuning optimizer optuna
import optuna

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(10,random_state=42)

## Model_4 : Linear Discriminant Analysis Classifiers 

### Computing the test set roc_auc score using Vanilla LDA classifier with SVD solver with no other hyperparameters to tune.

# Importing LDA Classifier
from sklearn.discriminant_analysis import LinearDiscriminantAnalysis

# Instantiating the classifier object 
lda_svd = LinearDiscriminantAnalysis(solver='svd')

# Fitting the classifier on the Training data directly as there are no hyperparameters to tune
lda_svd.fit(X_train_red,y_train)

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} is '.format(f_set, t_set, model_name), roc_auc_score(y, y_pred[:,1]))

# Calculating the Reduced feature training set roc_auc score 
cal_roc_auc(X_train_red, y_train, lda_svd, 'Reduced feature', 'training', 'LDA_SVD Classifier')

# Calculating the Reduced feature test set roc_auc score 
cal_roc_auc(X_test_red, y_test, lda_svd, 'Reduced feature', 'test', 'LDA_SVD Classifier')

### Computing the best hyperparameters for the Linear Discriminant Classifier with ‘eigen’ solver using Reduced Feature Training Set.

# Instantiating the classifier object 
lda_eigen = LinearDiscriminantAnalysis(solver='eigen')

# Defining the appropriate objective function for the LDA classifier
def objective_wrappper_lda(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Linear Discriminant Analysis parameters on the given training set X_tr, y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'shrinkage': trial.suggest_uniform('shrinkage', 0, 1)
        }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def train_test_roc_auc(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, lda_eigen, objective_wrappper_lda, cv_strat, n_trials=200)

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ',best_study_params)

# Obtaining the best reduced feature LDA  model with eigen solver by setting best study parameters.
lda_eigen = lda_eigen.set_params(**best_study_params)

# fitting the best LDA_eigen  model on the reduced feature training set
lda_eigen.fit(X_train_red, y_train)

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red ,y_train , lda_eigen, 'Reduced feature', 'training', 'LDA Classifier with eigen solver')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red ,y_test , lda_eigen, 'Reduced feature', 'test', 'LDA Classifier with eigen solver')

###  Computing the test set roc_auc score using LDA classifier with eigen solver & shrinkage utilizing the Ledoit-Wolf lemma with no other hyperparameters to tune.


# Instantiating the LDA classifier object using  eigen solver utilizing shrinkage using the Ledoit-Wolf lemma
lda_eigen_auto = LinearDiscriminantAnalysis(solver='eigen',shrinkage='auto')

# fitting the best LDA_eigen model with Ledoit-Wolf shrinkage on the reduced feature training set
lda_eigen_auto.fit(X_train_red, y_train)

# Calculating the Reduced feature training set roc_auc score using Ledoit-Wolf shrinkage
cal_roc_auc(X_train_red , y_train, lda_eigen_auto, 'Reduced feature', 'training', 'LDA Classifier with Ledoit-Wolf shrinkage')

# Calculating the Reduced feature test set roc_auc score using Ledoit-Wolf shrinkage
cal_roc_auc( X_test_red, y_test, lda_eigen_auto, 'Reduced feature', 'test', 'LDA Classifier with Ledoit-Wolf shrinkage')

# Saving the Reduced feature best LDA model 
import joblib
joblib.dump(lda_eigen,'Linear_Dis_Reduced.joblib')

### Calculating R_R ratio for Vanilla LDA Classifier using SVD solver

# Computing the CV scores using sklearn's cross_val_score
score_lda = cross_val_score(lda_svd, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the vanilla lda Classifier using roc_auc metric is: ',np.mean(score_lda))

print('The risk associated with the vanilla lda Classifier using roc_auc metric is: ',np.std(score_lda))

R_R_Ratio_lda = np.mean(score_lda)/np.std(score_lda)

print('The reward risk ratio for the vanilla lda Classifier using roc_auc metric is: ',R_R_Ratio_lda)

#### The R_R ratio for the vanilla lda Classifier using roc_auc metric is:  51.152886235836306

### Calculating R_R ratio for the best  lda Classifier using eigen solver & shrinkage

# Loading the best lda classifier with eigen solver & shrinkage model.
import joblib
lda_eigen = joblib.load('Linear_Dis_Reduced.joblib')

# Computing the CV scores using sklearn's cross_val_score
score_lda_eigen = cross_val_score(lda_eigen, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best lda classifier with eigen solver using roc_auc metric is: ',np.mean(score_lda_eigen))

print('The risk associated with the best lda classifier with eigen solver using roc_auc metric is: ',np.std(score_lda_eigen))

R_R_Ratio_lda_eigen = np.mean(score_lda_eigen)/np.std(score_lda_eigen)

print('The reward risk ratio for the  best lda classifier with eigen solver using roc_auc metric is: ',R_R_Ratio_lda_eigen)

#### The R_R ratio for the best lda classifier with eigen solver using roc_auc metric is:  50.589549888479354

### Calculating R_R ratio for the lda Classifier with eigen solver & shrinkage utilizing the Ledoit-Wolf lemma

# Computing the CV scores using sklearn's cross_val_score
score_lda_auto = cross_val_score(lda_eigen_auto, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: ',np.mean(score_lda_auto))

print('The risk associated with the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: ',np.std(score_lda_auto))

R_R_Ratio_lda_auto = np.mean(score_lda_auto)/np.std(score_lda_auto)

print('The reward risk ratio for the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: ',R_R_Ratio_lda_auto)

#### The R_R ratio for the lda Classifier utilizing LW lemma shrinkage using roc_auc metric is: 49.40997736726593

### R_R Ratio for the best LDA classifier using reduced feature set is:  51.152886235836306, corresponding to vanilla LDA Classifier.


## Observations:

### 1) The Best LDA model with eigen solver & tuned shrinkage performed  worse than the Logistic Regression Model as well as Tree based Bagging models on the test set. This was expected as the underlying Feature space is not multivariate normal and doesn't have the same covariance matrix for both the classes ,which is the underlying assumption of the LDA model. However its R_R ratio was greater than that of the logistic Regression & less than those of Tree based Bagging Classifiers.
### 2) Among all lda variants, the R_R ratio is greatest for plain vanilla lda classifier with SVD solver.


## Model_5 : Quadratic Discriminant Analysis Classifier

### Computing the Reduced Feature test set roc_auc score using QDA classifier 

from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis

# Instantiating the QDA classifier object 
QDA = QuadraticDiscriminantAnalysis()

# Defining the appropriate objective function for the QDA classifier
def objective_wrappper_qda(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Quadratic Discriminant Analysis parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'reg_param': trial.suggest_uniform('reg_param', 0, 1)
        }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, QDA, objective_wrappper_qda, cv_strat, n_trials=200)

print('The best roc_auc_score for the study is: ',best_study_score)

print('The best study parameters for the classifier are: ',best_study_params)

# Obtaining the best reduced feature QDA  model with eigen solver by setting best study parameters.
QDA = QDA.set_params(**best_study_params)

# fitting the best QDA  model on the reduced feature training set
QDA.fit(X_train_red,y_train)

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red , y_train, QDA, 'Reduced feature', 'training', 'QDA classifier')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, QDA, 'Reduced feature', 'test', 'QDA classifier')

# Saving the Reduced feature best QDA model 
joblib.dump(QDA,'Quadratic_Dis_Reduced.joblib')

### Calculating R_R ratio for best QDA Classifier.

# Loading the best QDA Classifier model
import joblib
QDA = joblib.load('Quadratic_Dis_Reduced.joblib')

# Computing the CV scores using sklearn's cross_val_score
score_qda = cross_val_score(QDA, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best QDA Classifier using roc_auc metric is: ',np.mean(score_qda))

print('The risk associated with the best QDA Classifier using roc_auc metric is: ',np.std(score_qda))

R_R_Ratio_qda = np.mean(score_qda)/np.std(score_qda)

print('The reward risk ratio for the best QDA Classifier using roc_auc metric is: ',R_R_Ratio_qda)

#### The R_R Ratio for the best QDA classifier using reduced feature set is:  38.5216883550956

## Observations:
### 1) The tuned QDA model has the worst test set roc_auc score as well as R_R ratio of all the fitted models till now. This was expected as the underlying Feature space is not multivariate normal, which is the underlying assumption of the QDA model. Any departure from normality affects the QDA more than the LDA, which is also observed here.
### 2) Owing to both very low R_R ratio and test set roc_auc, we can rule out the QDA model for this dataset.

### R_R Ratio for the best  classifier  in Discriminant Analysis family utilizing roc_auc metric is:  51.152886235836306,  corresponding to Vanilla LDA Classifier.

### _Thus keeping everything into account , for this dataset , the best discriminant classifier is LDA model with eigen solver & tuned shrinkage ._
