==> 7_a_Modeling.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)

# Importing Cross_val_score & Stratified K fold 
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import StratifiedKFold

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Loading both versions of Training sets
X_train_red = pd.read_csv('X_train_final.csv')
X_train = pd.read_csv('X_train_full_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading both versions of Test sets
X_test_red = pd.read_csv('X_test_final.csv')
X_test = pd.read_csv('X_test_full_final.csv')
y_test = pd.read_csv('y_test.final.csv')

## Model_6: Light Gbm classifier with Tuned Hyperparameters using Optuna.

# Importing Optuna and lightgbm 
from lightgbm import LGBMClassifier
import optuna

# Defining the appropriate objective function for the Light GBM classifier

def objective_wrappper_lgbm(X_tr, y_tr, cls=None):
    '''
    Optimizes classifier's cls (LightGBM here) parameters on the given training set X_tr, y_tr
    using cv_strat cross-validation object
    
    '''
    # Defing the class weights
    # Initializing the K Fold object
    cv_strat = StratifiedKFold(10,random_state=42)
    cl_weight = ['balanced',{0:1.0,1:9.0},{0:1.0,1:10},{0:1.0,1:11},{0:1.0,1:12},{0:1.0,1:13}]
    def objective(trial):
        params = {
        'class_weight': trial.suggest_categorical('class_weight',cl_weight),
        'n_estimators': trial.suggest_int('n_estimators', 10, 1500),
        'min_child_samples': trial.suggest_int('min_child_samples', 10, 250),
        'reg_alpha': trial.suggest_uniform('reg_alpha', 0, 10),
        'reg_lambda': trial.suggest_uniform('reg_lambda', 0, 10),
        'num_leaves': trial.suggest_int('num_leaves', 2, 300),
        'max_depth': trial.suggest_int('max_depth', 2, 100),
        'colsample_bytree': trial.suggest_uniform('colsample_bytree', 0.6, 1.0),
        'learning_rate': trial.suggest_loguniform('learning_rate', 1e-3, 1e-1)
        }
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def train_test_roc_auc(X_tr, y_tr, cls, obj_func, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier on the Training set and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Initializing the LightGBM Classifier
lgb_s = LGBMClassifier(random_state=42, objective='binary', n_jobs=5)

#### Computing the best hyperparameters for the LightGBM classifier using Full Feature Training Set, _just for trial sake_

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train, y_train, lgb_s, objective_wrappper_lgbm, n_trials=100)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ', best_study_params))

#### Computing the  full feature roc_auc score for the test data using the best study Parameters

# Obtaining the best full feature Light GBM model by setting best study parameters.
lgb_f = lgb_s.set_params(**best_study_params)

# fitting the best lightgbm model on the full feature training set
lgb_f.fit(X_train, y_train)

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} classifier is '.format(f_set,t_set,model_name),roc_auc_score(y, y_pred[:,1]))

# Calculating the full feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train, y_train, lgb_f, 'full feature', 'training', 'Light Gbm')

# Calculating the full feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test, y_test, lgb_f, 'full feature', 'test', 'Light Gbm')

#Saving the Full feature set best Light Gbm Classifier 
import joblib
joblib.dump(lgb_f,'Light_Gbm_Full.joblib')

#### Computing the best hyperparameters for the  Light Gbm Classifier using Reduced feature Training Set.

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, lgb_s, objective_wrappper_lgbm, n_trials=100)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ', best_study_params))

# Obtaining the best Reduced feature LightGbm model by setting best study parameters.
lgb_red = lgb_s.set_params(**best_study_params)

# fitting the best lightgbm model on the Reduced feature training set
lgb_red.fit(X_train_red, y_train)

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, lgb_red, 'Reduced feature', 'training', 'Light Gbm')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, lgb_red, 'Reduced feature', 'test','Light Gbm')

# Saving the Reduced feature set best Light Gbm Classifier 
joblib.dump(lgb_red,'Light_Gbm_Reduced.joblib')

### Calculating R_R ratio for best Light GBM Classifier trained on reduced feature training set.

# Loading the best Light GBM  Classifier model
import joblib
lgb_red = joblib.load('Light_Gbm_Reduced.joblib')

# Instantiating the Stratified fold object
cv_strat = StratifiedKFold(10,random_state=42)

# Computing the CV scores using sklearn's cross_val_score
score_Light_Gbm = cross_val_score(lgb_red, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best Light GMB Classifier using roc_auc metric is: ',np.mean(score_Light_Gbm))

print('The risk associated with the best Light GMB Classifier using roc_auc metric is: ',np.std(score_Light_Gbm))

R_R_Ratio_Light_Gbm = np.mean(score_Light_Gbm)/np.std(score_Light_Gbm)

print('The reward risk ratio for the best Light Gbm Classifier using roc_auc metric is: ',R_R_Ratio_Light_Gbm)

#### The R_R Ratio for the best Light Gbm classifier, trained on reduced feature training set, using roc_auc score is:  45.64661929591216

## Observations:
### 1) From the above analysis we can clearly see that reduced feature test set roc_auc score is more than full feature roc_auc score , thereby clearly substantiating  the presence of noisy predictors in the feature set , _as has been observed before._
### 2) The Light Gbm classifier has beautifully fitted the dataset with no indications of overfitting and of all the classifiers tested till now , has given the best test set roc_auc score. May be xgboost tested next may beat that.
### 3) The Light Gbm has poorer R_R ratio as compared to those of Tree based bagging classifiers, indicating too much roc_auc variability across CV folds.
### 4) Surprisingly R_R ratio for the Light GBM model is on the lower side of the spectrum.


## Model_7: XGboost Classifier with Tuned Hyperparameters using Optuna.

# Importing Xgboost Classifier
import xgboost as xgb

# Computing the ratio of -ve to +ve classes. Note here class_weight is a series and should be converted to float value
# using float() to be used as a sclar inside xgboost. 
class_weight = float((y_train==0).sum()/(y_train==1).sum())

# Defining the appropriate objective function for the XGboost classifier

def objective_wrappper_xgb(X_tr, y_tr, cls=None, cv=None):
    '''
    Optimizes classifier's cls (Xgboost here) parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    # Initializing the K Fold object
    cv_strat = StratifiedKFold(cv, random_state=42)
    
    def objective(trial):
        params = {
        'n_estimators': trial.suggest_int('n_estimators', 10, 1500),
        'reg_alpha': trial.suggest_uniform('reg_alpha', 0, 10),
        'reg_lambda': trial.suggest_uniform('reg_lambda', 0, 10),
        'max_delta_step': trial.suggest_int('max_delta_step', 0, 10),
        'max_depth': trial.suggest_int('max_depth', 2, 100),
        'colsample_bytree': trial.suggest_uniform('colsample_bytree', 0.6, 1.0),
        'learning_rate': trial.suggest_loguniform('learning_rate', 1e-3, 1e-1)
        }
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Defining the evaluation function for study's best parameters
def train_test_roc_auc(X_tr, y_tr, cls, obj_func, cv=None, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier on the Training set and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Instantiating the xgboost classifier  
xgb_s = xgb.XGBClassifier(random_state=42,n_jobs=5,objective='binary:logistic',scale_pos_weight=class_weight)

#### Computing the best hyperparameters for the  XGboost Classifier using Reduced feature Training Set.

# Importing Optuna 
import optuna

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_red, y_train, xgb_s, objective_wrappper_xgb, cv=10)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ',best_study_params))

# Obtaining the best Reduced feature Xgboost model by setting best study parameters.
xgb_red = xgb_s.set_params(**best_study_params)

# fitting the best Xgboost model on the Reduced feature training set
xgb_red.fit(X_train_red, y_train)

# Calculating the Reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, xgb_red, 'Reduced feature', 'training', 'Xgboost')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, xgb_red, 'Reduced feature', 'test', 'Xgboost')

# Saving the Reduced feature set best Xgboost Classifier 
import joblib
joblib.dump(xgb_red, 'XGB_Reduced.joblib')

### Calculating R_R ratio for best Xgboost Classifier.

# Loading the best Xgboost  Classifier model
xgb_red = joblib.load('XGB_Reduced.joblib')

# Computing the CV scores using sklearn's cross_val_score
score_xgb = cross_val_score(xgb_red, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best Xgboost Classifier using roc_auc metric is: ',np.mean(score_xgb))

print('The risk associated with the best Xgboost Classifier using roc_auc metric is: ',np.std(score_xgb))

R_R_Ratio_Xgboost = np.mean(score_xgb)/np.std(score_xgb)

print('The reward risk ratio for the best Xgboost Classifier using roc_auc metric is: ',R_R_Ratio_Xgboost)

#### The R_R Ratio for the best XgBoost classifier using roc_auc score is: 48.43484617872931

## Observations:
### 1) Xgboost model ,conceptually similar to Light GBM , wonderfully fitted this dataset , with no apparent signs of overfitting. But surprisingly the test set roc_auc score for tuned Xgboost is less than that of Light GBM. 
### 2) Xgboost has a higher R_R ratio than that of Light GBM & hence should be preferred model of choice based on Rewards Risk ratio  metric, but has a higher computational cost than that of Light Gbm.
### 3) Both boosting models have been outshone by Logistic Regression & Tree based Bagging models in the domain of R_R ratio metric, due to their comparatively low roc_auc score Std Dev.

### R_R Ratio for the best Tree based boosting classifier using roc_auc score is:  48.43484617872931 , corresponding to tuned XgBoost classifier.

### _Thus keeping everything into account (including the computational costs) , for this dataset , the best Tree based boosting classifier is tuned Light Gbm classifier._
