==> 8_a_Modeling.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)

# Importing Cross_val_score & Stratified K fold 
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import StratifiedKFold

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Loading Reduced feature Training set
X_train_red = pd.read_csv('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading Reduced feature Test set
X_test_red = pd.read_csv('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

## Model_8: SVM Classifier with Tuned Hyperparameters using Random Search CV.

# Importing Random Search CV objects
from sklearn.model_selection import RandomizedSearchCV

# Initializing the K Fold object with 5 folds instead of 10 as SVM is computationally expensive.
cv_strat = StratifiedKFold(5,random_state=42)

# Formulating the hyperparameter grid for SVM Classifier
params_svm = {'C': [.001,.01,.1,.5,1,2,4,6,8,10,20,40,60,80,100],
              'gamma':  [.0001,.001,.01,.1,.5,1,2,4,6,8,10]+['scale','auto'],
              'class_weight':['balanced',{0:1.0,1:9.0},{0:1.0,1:10},{0:1.0,1:11},{0:1.0,1:12}]
             }
              

# Importing the SVM Classifier
from sklearn.svm import SVC

# Instantiating the SVM Classifier
svm_s = SVC(probability=True,cache_size=500,random_state=42,kernel='rbf')

# Initializing Random_search CV object
randcv_svm = RandomizedSearchCV(svm_s,param_distributions=params_svm,scoring='roc_auc',n_jobs=5,cv=cv_strat,
                              random_state=42,n_iter=50)

# Running Random Search CV on Training set
randcv_svm.fit(X_train_red,y_train)

# Getting the parameters of the best SVM Model
randcv_svm.best_params_

# Getting the best roc_auc 5 Fold Score 
randcv_svm.best_score_

# Getting the best SVM estimator from Random search CV, which has already been trained on the entire Training set
svm_red = randcv_svm.best_estimator_

# Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the best {} classifier is '.format(f_set, t_set, model_name),roc_auc_score(y, y_pred[:,1]))

# Calculating the Reduced feature training set roc_auc score using the best SVM model
cal_roc_auc(X_train_red, y_train, svm_red, 'Reduced feature', 'training', 'SVM')

# Calculating the Reduced feature test set roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, svm_red, 'Reduced feature', 'test', 'SVM')

# Saving the Reduced feature set best SVM Classifier 
import joblib
joblib.dump(svm_red,'SVM_Reduced.joblib')

### Calculating R_R ratio for best SVM Classifier.

# Loading the best SVM  Classifier model
import joblib
svm_red = joblib.load('SVM_Reduced.joblib')

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(10,random_state=42)

# Computing the CV scores using sklearn's cross_val_score
score_svm = cross_val_score(svm_red, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the best SVM Classifier using roc_auc metric is: ',np.mean(score_svm))

print('The risk associated with the best SVM Classifier using roc_auc metric is: ',np.std(score_svm))

R_R_Ratio_svm = np.mean(score_svm)/np.std(score_svm)

print('The reward risk ratio for the best SVM Classifier using roc_auc metric is: ',R_R_Ratio_svm)

#### The R_R Ratio for the best SVM classifier using roc_auc score is: 47.16045814402432

## Observations:
### 1) SVM classifier shows no sign of overfitting  as both training and test roc_auc scores are almost equal to each other . But both training & test roc_auc scores for best SVM Model  are less than the corresponding scores from the best logistics Regression model( which belongs to the same classifier family  as SVM).
### 2) SVM fits really really slowly & is only a feasible option for small or medium sized datasets.
### 3) The R_R ratio  for the best SVM model is less than that of best Logistic Regression, which has much lower computational complexity. _Thus for this dataset, Logistic Regression outperforms SVM on all fronts._
