==> 11_a_Modeling.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)

# Loading Reduced feature Training set
X_train_red = pd.read_csv('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading Reduced feature Test set
X_test_red = pd.read_csv('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

## Loading all the best models from various categories

# Importing Joblib module
import joblib

import tensorflow

from tensorflow import keras

# Importing best Logistic regression Classifier
lr = joblib.load('Log_Reg_Reduced.joblib')

# Importing best Random Forest Classifier
rf = joblib.load('Rand_Forest_Reduced.joblib')

# Importing best Light Gbm Classifier
lgbm = joblib.load('Light_Gbm_Reduced.joblib')

# Importing best Neural Net Classifier
neural = keras.models.load_model('Best_model_Selu_eq_Learn.h5')

# Importing best Linear Discriminant Analysis Classifier
lda = joblib.load('Linear_Dis_Reduced.joblib')

## Model_13: Voting Classifier with Default Parameters & Soft_Voting.

# Importing Voting classifier from sklearn
from sklearn.ensemble import VotingClassifier

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

voting_clf = VotingClassifier(estimators=[('logistic_Reg',lr),('Random_Forest',rf),('Light_Gbm',lgbm),
                                           ('Linear_Dis',lda)],voting='soft',n_jobs=5)

# Fitting the vanilla voting classifier on the Reduced Feature Training set
voting_clf.fit(X_train_red, y_train)

#Defining the function to calculate the roc_auc score for the feature sets
def cal_roc_auc(X, y, cls, f_set, t_set, model_name):
    ''' Calculates the roc auc score using the best study parameters 
        f_set : String: specifies 'full feature', 'Reduced feature'
        t_set: String: specifies 'training', 'test'
        model_name: String: specifies Name of the model '''
        
    y_pred = cls.predict_proba(X)
    print('The roc_auc_score for the {} {} set using the {} is '.format(f_set,t_set,model_name),roc_auc_score(y,y_pred[:,1]))

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, voting_clf,'reduced feature','training',' default Voting Classifier')

# Calculating the reduced feature test roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, voting_clf,'reduced feature','test','default Voting Classifier')

# Saving the default voting Classifier
import joblib
joblib.dump(voting_clf,'Voting_Red_default.joblib')

### Calculating R_R ratio for default Voting Classifier.

# Importing required Libraries
from sklearn.model_selection import StratifiedKFold
from sklearn.model_selection import cross_val_score

# Instantiating the Stratified K fold object
cv_strat = StratifiedKFold(10,random_state=42)

# Computing the CV scores using sklearn's cross_val_score
score_voting_default = cross_val_score(voting_clf, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the default Voting Classifier using roc_auc metric is: ',np.mean(score_voting_default))

print('The risk associated with the default Voting Classifier using roc_auc metric is: ',np.std(score_voting_default))

R_R_Ratio_voting_default = np.mean(score_voting_default)/np.std(score_voting_default)

print('The reward risk ratio for the default Voting Classifier using roc_auc metric is: ',R_R_Ratio_voting_default)

#### R_R Ratio for the default Voting classifier using reduced feature set is: 49.271728792879195

## Observations: 
### 1) The test set roc_auc score for the default voting classifier is more than that of component Logistic Regression & LDA, but less than that of  Random Forest & LightGBM. May be changing the weights of voting classifier result in better results.
### 2) On the other hand, the R_R ratio for the default voting classifier model is more than that of Logistic regression & LightGBM , but less than that of Random forest & LDA. So clearly Random Forest bests  default Voting classifier hands down.

## Model_13: Voting Classifier with Tuned weights.

# Instantiating a new voting classifier object
voting_clf_2 = VotingClassifier(estimators=[('logistic_Reg',lr),('Random_Forest',rf),('Light_Gbm',lgbm),
                                           ('Linear_Dis',lda)],voting='soft',n_jobs=5)

# Defining the appropriate objective function for the best weights of Voting classifier.
def objective_wrappper_Vt_2(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes voting classifier (cls) weights parameter on the given training set X_tr,y_tr
      
    '''
    
    def objective(trial):
        w1 = trial.suggest_uniform('w1',-1,1)
        w2 = trial.suggest_uniform('w2',-1,1)
        w3 = trial.suggest_uniform('w3',-1,1)
        w4 = trial.suggest_uniform('w4',-1,1)
        
        params = {
            'weights':[w1,w2,w3,w4]                 
            }
        
        cls.set_params(**params)#Initializing the model with the parameter
               
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))
    
    return objective

#### Computing the best weights for the Voting Classifier_2 using Optuna.

# Defining the evaluation function for study's best parameters
def train_test_roc_auc(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Extracting the best model parameters and best study score for the 2nd Voting Classifier
best_study_score,best_study_params = train_roc_auc(X_train_red, y_train, voting_clf_2, objective_wrappper_Vt_2,
                                                   cv_strat)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ',best_study_params))

# Obtaining the best tuned voting clasiifier model by setting best study parameters.
voting_clf_2 = voting_clf_2.set_params(weights=[best_study_params['w1'],best_study_params['w2'],
                                                 best_study_params['w3'],best_study_params['w4']])

# fitting the best Voting Classifier on the reduced feature training set
voting_clf_2.fit(X_train_red, y_train)

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, voting_clf_2, 'reduced feature', 'training', 'Voting Classifier')

# Calculating the reduced feature test roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, voting_clf_2, 'reduced feature', 'test', 'Voting Classifier')

# Saving the best voting Classifier
import joblib
joblib.dump(voting_clf_2,'Voting_Red.joblib')

### Calculating R_R ratio for the tuned Voting Classifier.

# Loading the best SVM  Classifier model
import joblib
voting_clf_2 = joblib.load('Voting_Red.joblib')

# Computing the CV scores using sklearn's cross_val_score
score_voting_best = cross_val_score(voting_clf_2, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the tuned Voting Classifier using roc_auc metric is: ',np.mean(score_voting_best))

print('The risk associated with the tuned Voting Classifier using roc_auc metric is: ',np.std(score_voting_best))

R_R_Ratio_voting = np.mean(score_voting_best)/np.std(score_voting_best)

print('The reward risk ratio for the tuned Voting Classifier using roc_auc metric is: ',R_R_Ratio_voting)

#### The R_R Ratio for the tuned Voting classifier using roc_auc metric is 47.28444783471113

## Observations: 
### 1)  The test set roc_auc score of the tuned voting classifier is more than that of all component models (by a good margin ), but for that of Light  GBM and also exceeds that of untuned  default voting classifier (voting_clf)
### 2) On the other hand, quite surprisingly the R_R ratio for the tuned voting classifier model is more than that of only Light GBM (among the component models) and is also beaten down by that of the default voting classifier.  
### 3) Again after accounting for all the factors, the tuned Random forest classifier still reigns supreme. 

## Model_14: Voting Classifier with Tuned weights without LDA.

voting_clf_3 = VotingClassifier(estimators=[('logistic_Reg',lr),('Random_Forest',rf),('Light_Gbm',lgbm)],
                                           voting='soft',n_jobs=5)

# Defining the appropriate objective function for the best weights of Voting classifier classifier
def objective_wrappper_Vt_3(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes voting classifier (cls) weights parameter on the given training set X_tr, y_tr
      
    '''
    
    def objective(trial):
        
        w1 = trial.suggest_uniform('w1',0,1)
        w2 = trial.suggest_uniform('w2',0,1)
        w3 = trial.suggest_uniform('w3',0,1)
        
        
        params = {
            'weights':[w1,w2,w3]                 
            }
        
        cls.set_params(**params)#Initializing the model with the parameter
               
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))
    
    return objective

# Defining the evaluation function for study's best parameters
def train_roc_auc(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_roc_auc(X_train_red, y_train, voting_clf_3, objective_wrappper_Vt_3,
                                                   cv_strat)

print('The best roc_auc_score for the study is: ',best_study_score)

print(('The best study parameters for the classifier are: ',best_study_params))

# Obtaining the best reduced feature LR model by setting best study parameters.
voting_clf_3 = voting_clf_3.set_params(weights=[best_study_params['w1'],best_study_params['w2'],
                                                 best_study_params['w3']])

# fitting the best Voting Classifier on the full feature training set
voting_clf_3.fit(X_train_red, y_train)

# Saving the best voting Classifier without lda
import joblib
voting_clf_3 = joblib.load('Voting_Red_without_lda.joblib')

# Calculating the reduced feature training set roc_auc score using the best study parameters
cal_roc_auc(X_train_red, y_train, voting_clf_3, 'reduced feature', 'training', 'tuned Voting Classifier without lda')

# Calculating the reduced feature test roc_auc score using the best study parameters
cal_roc_auc(X_test_red, y_test, voting_clf_3, 'reduced feature', 'test', 'tuned Voting Classifier without lda')

# Saving the best voting Classifier without lda
import joblib
joblib.dump(voting_clf_3,'Voting_Red_without_lda.joblib')

### Calculating R_R ratio for the tuned Voting Classifier without  LDA

voting_clf_3

# Computing the CV scores using sklearn's cross_val_score
score_voting_3 = cross_val_score(voting_clf_3, X_train_red, y_train, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the tuned Voting Classifier without lda using roc_auc metric is: ',np.mean(score_voting_3))

print('The risk associated with the tuned Voting Classifier using roc_auc metric is: ',np.std(score_voting_3))

R_R_Ratio_voting_without_lda = np.mean(score_voting_3)/np.std(score_voting_3)

print('The reward risk ratio for the tuned Voting Classifier without lda using roc_auc metric is: ',R_R_Ratio_voting_without_lda)

#### The R_R Ratio for the tuned Voting classifier without lda using roc_auc metric is 48.19861295727324

## Observations: 
### 1) The voting classifier with lda outperformed the one without lda, in terms of test set roc_auc score, which was expected as the voting classifier performs well when the constituent models are many and well diversified.
### 2)  The test set roc_auc score of the tuned voting classifier without lda is more than that of all component models (by a good margin ), but for that of Light  GBM and also exceeds that of untuned  default voting classifier (voting_clf)
### 3) Again, the R-R ratio for the tuned voting classifier without lda is more than that of only Light GBM (among the component models) & tuned voting classifier with lda. But is beaten down by that of the default voting classifier. 
### 4) Taking everything into consideration( such as computational cost), the tuned Random forest classifier has beaten all the voting classifiers & is clearly the winner till here.


### Best R_R Ratio for the voting classifier family using roc_auc metric is:  49.271728792879195 ,corresponding to default Voting Classifier

### _Thus keeping everything into account (including the computational costs), for this dataset , the best Voting Classifier is the tuned Voting Classifier with lda._
