==> 13.1_a_Modeling.ipynb <==
# Importing Required Python Packages
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
pd.set_option('display.max_columns',None)

# Loading Reduced feature Training set
X_train_red = pd.read_csv('X_train_final.csv')
y_train = pd.read_csv('y_train.final.csv')

# Loading Reduced feature Test set
X_test_red = pd.read_csv('X_test_final.csv')
y_test = pd.read_csv('y_test.final.csv')

## Model_16: Stacking Classifier

####  Loading  the best Voting Classifier model & Neural Network (with Equal Nodes in all the Layers)

# Importing Joblib module
import joblib

import tensorflow
from tensorflow import keras

# Importing best Voting Classifier
voting_clf = joblib.load('Voting_Red.joblib')

### Preparing the Training Set for Stacking Classifier

# Instantiating an empty DataFrame to store training set for Stacking Classifier
df_train_stack = pd.DataFrame(columns=['Voting','Neural','True'])

# Initializing the 10 Fold object
from sklearn.model_selection import StratifiedKFold
cv_strat = StratifiedKFold(10,random_state=42)

#### Getting predictions of the Voting & Neural Net classifier for each of 10 folds of Training set.

# Importing the Sklearn's roc_auc_score module
from sklearn.metrics import roc_auc_score

# Importing cross val score from sklearn
from sklearn.model_selection import cross_val_score

# Importing train test split from Sklearn to produce validation set
from sklearn.model_selection import train_test_split

# Defining the exponential decay learning rate.
def exponential_decay_fn(epoch):
            return 0.01 * 0.1**(epoch /4 )

def Neural_Pred(clf, X_tr, y_tr, X_tst):
    ''' 
    This function returns the predictions of Neural Classifier(clf) for the test folds 
    Parameters:
    X_tr : Training set Features
    y_tr : Training set Labels
    X_tst : Test set Features
    '''
    # Splitting the Training set further into training & validation set.
    X_tr_r, X_val, y_tr_r, y_val = train_test_split(X_tr, y_tr, test_size=0.1, random_state=42, stratify=y_tr)
    
    # Compiling & Training the Neural Net
    clf.compile(loss="binary_crossentropy", optimizer=keras.optimizers.Nadam(beta_1=0.9, beta_2=0.999), metrics=["accuracy"])
    
    # defining Checkpoints
    checkpoint_cb = keras.callbacks.ModelCheckpoint('best_model.h5',save_best_only=True) # 1st Callback
    early_stopping_cb = keras.callbacks.EarlyStopping(patience=10)# 2nd Callback
    lr_scheduler_cb = keras.callbacks.LearningRateScheduler(exponential_decay_fn)# 3rd Callback
    
    # Fitting The model
    clf.fit(X_tr_r, y_tr_r, epochs=50, validation_data=(X_val,y_val),batch_size=32,
    class_weight={0: 1.0, 1: 10.0},callbacks=[checkpoint_cb,early_stopping_cb,lr_scheduler_cb])
    
    # Loading the best Neural model after training & before making predictions
    clf_best = keras.models.load_model('best_model.h5')    
    
    # Saving the predictions for every fold in a numpy array
    return (clf_best.predict_proba(X_tst).flatten())

# Computing Voting Classifier's & Neural Net 10 fold conditional prob predictions on the training set for Stacking classifier
for train_index, test_index in cv_strat.split(X_train_red, y_train):
    # Creating  Folds
    X_tr, X_tst = X_train_red.loc[train_index], X_train_red.loc[test_index]
    y_tr, y_tst = y_train.loc[train_index], y_train.loc[test_index]
    
    # Fitting the best Voting Classifier on the Training Folds
    voting_clf.fit(X_tr, y_tr)
    # Making Predictions on the testing Fold.
    y_pred_voting = voting_clf.predict_proba(X_tst)[:,1]
    
    
    # Loading the fresh best Neural net classifier
    neural = keras.models.load_model('Best_model_Selu_eq_Learn.h5')
    # Fitting the best Neural Net on Training Folds & obtaining  predictions
    y_pred_neural = Neural_Pred(neural, X_tr, y_tr, X_tst)
    
    # Storing the predictions in a dataframe
    df_temp = pd.DataFrame(columns=['Voting','Neural','True'])
    df_temp['Voting'] = y_pred_voting
    df_temp['Neural'] = y_pred_neural
    df_temp['True'] = y_tst.values.flatten() # As y_tst is a dataframe we first convert it into a 2d numpy vector
    # array, which is is then converted to 1d array using .flatten()
    
    # Appending the df_temp to df_train_stack
    df_train_stack = df_train_stack.append(df_temp,ignore_index=True)   
    

# Getting the info of the Training dataframe of the  Stacking Classifier
df_train_stack.info()

# Converting the True Column to the int type
df_train_stack['True'] = df_train_stack['True'].astype('int8')

# Re-checking the info of the df_train_stack
df_train_stack.info()

# Printing the top 5 rows of the df_train_stack
df_train_stack.head()

### Selecting Random forest as the Stacking classifier as it has one of highest  test set roc_auc as well as R_R ratio for this dataset.

### Tuning the hyperparameters of the Random Forest on the Stacked Training set using Optuna

# Defining the class weights
cl_weight = [None,'balanced',{0:1.0,1:9.0},{0:1.0,1:10},{0:1.0,1:11},{0:1.0,1:12},{0:1.0,1:13},{0:1.0,1:14},{0:1.0,1:15}]

# Defining the appropriate objective function for the Random Forest classifier
def objective_wrappper_rf(X_tr, y_tr, cls=None, cv_strat=None):
    '''
    Optimizes Random Forest parameters on the given training set X_tr,y_tr
    using cv_strat cross-validation object
    
    '''
    
    def objective(trial):
        params = {
        'max_depth': trial.suggest_categorical('max_depth',list(range(2,50))+ [None]),
        'n_estimators':trial.suggest_int('n_estimators',100,2000,10),
        'class_weight':trial.suggest_categorical('class_weight',cl_weight),
        'min_samples_leaf':trial.suggest_loguniform('min_samples_leaf',.00001,.1)
          }
        
        cls.set_params(**params)#Initializing the model with the parameters 
    
        return np.mean(cross_val_score(cls, X_tr, y_tr, cv=cv_strat, n_jobs=5, scoring='roc_auc'))  
    return objective

# Importing  hyperparamater tuning optimizer optuna
import optuna

# Defining the evaluation function for study's best parameters
def train_test_roc_auc(X_tr, y_tr, cls, obj_func, cv_strat, n_trials=100):
    ''' Computes the best hyper parameters of the classsifier and returns 
    Optuna's study's best score & clasifier parameters'''
    study = optuna.create_study(direction='maximize')
    study.optimize(obj_func(X_tr, y_tr, cls, cv_strat), n_trials)
    best_score = study.best_value
    best_params = study.best_params
    return (best_score,best_params)


# Importing Random Forest Classifier from  Sklearn
from sklearn.ensemble import RandomForestClassifier

# Instantiating the Random forest classifier
rf_s = RandomForestClassifier(n_jobs=5, random_state=42)

# Segregating the Features and class labels from the Training Dataset
X_train_stacked = df_train_stack[['Voting','Neural']]
y_train_stacked = df_train_stack['True']

# Extracting the best model parameters and best study score
best_study_score,best_study_params = train_test_roc_auc(X_train_stacked, y_train_stacked, rf_s, objective_wrappper_rf, cv_strat, n_trials=150)

print('The best roc_auc_score for the study is: ', best_study_score)

print('The best study parameters for the classifier are: ', best_study_params)

# Obtaining the best Stacking Random Forest model by setting best study parameters.
rf_stack = rf_s.set_params(**best_study_params)

# fitting the best Stacking Random Forest model on the whole training set
rf_stack.fit(X_train_stacked, y_train_stacked)

### Preparing the Test set observations for the Stacking Classifier.

# Training the Voting Classifier on the whole Training set
voting_clf.fit(X_train_red,y_train)

# Computing the predictions of the voting classifier on the test set.
y_pred_test_voting = voting_clf.predict_proba(X_test_red)[:,1]

# Reloading the best Neural Classifier
neural = keras.models.load_model('Best_model_Selu_eq_Learn.h5')

# Computing the predictions of the Neural Classifier on the test set.
y_pred_test_neural = neural.predict_proba(X_test_red).flatten()

# Preparing the Test feature set for the Stacking Classifier
X_test_stacked = pd.DataFrame({'Voting':y_pred_test_voting,'Neural':y_pred_test_neural})

# Checking the info of the Stacking Test Features
X_test_stacked.info()

# Getting the Stacking Classifier Predictions for the Test set
y_pred_stacked = rf_stack.predict_proba(X_test_stacked)[:,1]

# Getting the stacking Clasiifier roc_auc score for the Test Set.
print('The test set roc_auc score for the Stacking Classifier is: ',roc_auc_score(y_test,y_pred_stacked))

### Calculating R_R ratio for Stacking Classifier.

# Computing the CV scores using sklearn's cross_val_score
score_Stacking = cross_val_score(rf_stack, X_train_stacked, y_train_stacked, cv=cv_strat, n_jobs=5, scoring='roc_auc')

print('The reward associated with the tuned Stacking Classifier using roc_auc metric is: ',np.mean(score_Stacking))

print('The risk associated with the tuned Stacking Classifier  using roc_auc metric is: ',np.std(score_Stacking))

R_R_Ratio_Stacking = np.mean(score_Stacking)/np.std(score_Stacking)

print('The reward risk ratio for the tuned Stacking Classifier using roc_auc metric is: ',R_R_Ratio_Stacking)

## Observations: 
### 1) The test set roc_auc score for the Stacking classifier is more than that of  Neural Net classifier , but less than that of the Voting classifier , both of which were used to create the training  as well as test set for the stacking classifier.
### 2) Similarly the R_R ratio of the Stacking classifier is approx. equal (although more) to that of the Voting Classifier, but much less than that of the Neural Net. Thus, even with added complexity, the Stacking Classifier still hasn't been able to beat the tuned Random Forest Classifier on this dataset.




###  The R_R Ratio for the tuned Stacking Classifier using roc_auc metric is:  48.351649337461005
